Fully refactor settings modal to use popover API Also split out some components that were making the file too long
